Changing project structure for the pages (home, Remote, About us, Contact)
