+++ src/pages/about/sections/AboutUsTitle.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SectionTitle from '../SectionTitle';
+import SectionTitle from '../../../components/SectionTitle';
 
 const AboutUsTitle: React.FC = () => {
     return (
