+++ src/pages/remote/RemoteWork.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import RemoteWorkTitle from '../components/remote/RemoteWorkTitle';
-import RemoteFeaturesSection from '../components/remote/RemoteFeaturesSection';
-import RemoteWorkExplain from "../components/remote/RemoteWorkExplain";
+import RemoteWorkTitle from './sections/RemoteWorkTitle';
+import RemoteFeaturesSection from './sections/RemoteFeaturesSection';
+import RemoteWorkExplain from "./sections/RemoteWorkExplain";
 const RemoteWork: React.FC = () => {
     return (
         <div className="container mx-auto mt-16">
