+++ src/pages/remote/sections/RemoteWorkTitle.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SectionTitle from "../SectionTitle";
+import SectionTitle from "../../../components/SectionTitle";
 
 const RemoteWorkTitle: React.FC = () => {
     return (
